featt(filter): Move to sub module
