+++ cg/store/api/findbasicdata.py
@@ -19,7 +19,7 @@ from cg.store.models import (
 from cg.store.api.base import BaseHandler
 from cg.store.filters.status_bed_filters import apply_bed_filter, BedFilters
 from cg.store.filters.status_bed_version_filters import BedVersionFilters, apply_bed_version_filter
-from cg.store.status_customer_filters import apply_customer_filter, CustomerFilter
+from cg.store.filters.status_customer_filters import apply_customer_filter, CustomerFilter
 from cg.store.filters.status_collaboration_filters import (
     CollaborationFilters,
     apply_collaboration_version_filter,
